pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "textual-synopsis"
version = "0.1.0"
description = "Text Alignment Tool"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.20",
    "pandas",
    "openpyxl",
    "biopython>=1.80",
    "numpy",
    "watchdog"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["textual_synopsis"]
